Index alias and file creation fixes

+++ include/http/Uri.hpp
@@ -34,10 +34,14 @@ class Uri {
   std::string getPath() const;
   std::string getQuery() const;
 
-  // Tries to decode the URI
+  // Tries to percent decode the URI
   // @return 0 if successful, non-zero otherwise
-  size_t decode();
+  int decode();
+
+  // Percent encodes an uri
   std::string encode() const;
+
+  // Generates the uri without percent encoding
   std::string generate() const;
 
   bool compare(const Uri &rhs) const;
